chore: migrate limit orders to jumper-backend REST client
Replace the generated limit-order client/types (gen-limit-order-api.sh,
limitOrderClient.ts, jumper-limit-order.ts) with the shared jumper-backend
client and types, and drop the now-unused NEXT_PUBLIC_LIMIT_ORDER_BACKEND_URL
env var. Also fixes a chainId mismatch when signing cancellation typed data,
and makes the repeat/modify order modals close (and reset) on wallet address
change, matching the cancel order modal.
