+++ src/components/LimitOrders/OrdersSection/OrdersTable.tsx
@@ -1,19 +1,25 @@
 'use client';
 
-import { useState } from 'react';
 import { useMediaQuery } from '@mui/material';
 import { DataTable } from '@/components/composite/DataTable/DataTable';
 import {
   Pagination,
   PaginationVariant,
 } from '@/components/core/Pagination/Pagination';
-import { type Order } from '@/types/jumper-limit-order';
+import { type LimitOrder as Order } from '@/types/jumper-backend';
 import { useOrderColumns } from './hooks';
 
-const PAGE_SIZE = 10;
-
 interface OrdersTableProps {
+  /** Already the current page's rows — pagination is server-side (offset/limit). */
   orders: Order[];
+  /** Total order count across all pages, from the backend's pagination meta. */
+  total: number;
+  /** The limit the backend actually applied to this page's request. */
+  pageSize: number;
+  /** Math.ceil(total / pageSize), computed once alongside the query result. */
+  pageCount: number;
+  page: number;
+  setPage: (page: number) => void;
   isLoading?: boolean;
   isExpanded?: boolean;
   stickyHeader?: boolean;
@@ -21,6 +27,11 @@ interface OrdersTableProps {
 
 export const OrdersTable = ({
   orders,
+  total,
+  pageSize,
+  pageCount,
+  page,
+  setPage,
   isLoading = false,
   isExpanded = false,
   stickyHeader = false,
@@ -29,7 +40,6 @@ export const OrdersTable = ({
   const isMobile = useMediaQuery((theme) => theme.breakpoints.down('sm'));
   const isTablet = useMediaQuery((theme) => theme.breakpoints.down('md'));
   const isSmallScreen = useMediaQuery((theme) => theme.breakpoints.down('lg'));
-  const [page, setPage] = useState(0);
 
   const maxVisiblePages = isMobile
     ? 1
@@ -39,14 +49,10 @@ export const OrdersTable = ({
         ? 3
         : 5;
 
-  // TODO: replace with server-side pagination once the API exposes limit/offset total count
-  const pageCount = Math.ceil(orders.length / PAGE_SIZE);
-  const pagedOrders = orders.slice(page * PAGE_SIZE, (page + 1) * PAGE_SIZE);
-
   return (
     <>
       <DataTable
-        rows={pagedOrders}
+        rows={orders}
         loading={isLoading}
         columns={columns}
         getRowKey={(order) => order.orderId}
@@ -62,9 +68,9 @@ export const OrdersTable = ({
           setPage={setPage}
           pagination={{
             page,
-            pageSize: PAGE_SIZE,
+            pageSize,
             pageCount,
-            total: orders.length,
+            total,
           }}
           maxVisiblePages={maxVisiblePages}
           sx={{ mt: 1, gap: (theme) => theme.spacing(0.5) }}
